reorg and clean up oci-runtime types
