+++ containrs/oci-runtime/src/v1.rs
@@ -3,3 +3,5 @@ mod state;
 
 pub use spec::*;
 pub use state::*;
+
+pub const VERSION: &str = "1.0.0";
